# code/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Instituto LG — Biblioteca de aulas API
  version: 0.7.0
  description: |
    API do painel (coordenação). Aluno não usa esta API — assiste na Eduq.
    Health: 200 com checks.database=ok; 503 se o banco cair (smoke de deploy).
    Auth Sanctum Bearer + Policy por recurso (conta `ativo`). Envio: PUT `/envios/{token}` (sem Bearer).
    Player público (web, não JSON): `GET /assistir/{token_publico}` (iframe) e mídia com URL assinada temporária.
    Cópia para a pasta compartilhada: job na fila `biblioteca` (aula pode estar Pronta com cópia pendente).
    Rotas privadas: 401 sem token; 403 se a conta não puder gerir o acervo.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /cursos/{curso}:
    parameters:
      - $ref: '#/components/parameters/CursoId'
    get:
      operationId: verCurso
      tags: [Cursos]
      responses:
        '200':
          description: Curso
        '404':
          description: Não encontrado
    put:
      operationId: atualizarCurso
      tags: [Cursos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
      responses:
        '200':
          description: Atualizado
    delete:
      operationId: excluirCurso
      tags: [Cursos]
      responses:
        '200':
          description: Removido
        '409':
          description: Tem turmas
...
  /turmas/{turma}:
    parameters:
      - $ref: '#/components/parameters/TurmaId'
    get:
      operationId: verTurma
      tags: [Turmas]
      responses:
        '200':
          description: Turma
    put:
      operationId: atualizarTurma
      tags: [Turmas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
      responses:
        '200':
          description: Atualizada
    delete:
      operationId: excluirTurma
      tags: [Turmas]
      responses:
        '200':
          description: Removida
        '409':
          description: Tem disciplinas
...
  /disciplinas/{disciplina}:
    parameters:
      - $ref: '#/components/parameters/DisciplinaId'
    get:
      operationId: verDisciplina
      tags: [Disciplinas]
      responses:
        '200':
          description: Disciplina
    put:
      operationId: atualizarDisciplina
      tags: [Disciplinas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
      responses:
        '200':
          description: Atualizada
    delete:
      operationId: excluirDisciplina
      tags: [Disciplinas]
      responses:
        '200':
          description: Removida
        '409':
          description: Tem aulas
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    CursoId:
      name: curso
      in: path
      required: true
      schema:
        type: integer
    TurmaId:
      name: turma
      in: path
      required: true
      schema:
        type: integer
    DisciplinaId:
      name: disciplina
      in: path
      required: true
      schema:
        type: integer
...